[FEAT] Introduce multiple sections
Introduce the ability to play multi parallel sections.
The evaluation state is now stored in the SectionEvaluator, instead of
on the ast.

Type and Evaluation information are now stored in `extras` in the ast
node.

Rework `Player` to use a linked list, instead of the `Mixer` with arena
from Zynth.